Moved tests to I18nIT.java
