+++ vaadin-login-flow/src/main/resources/i18n.json
@@ -12,7 +12,7 @@
   },
   "errorMessage": {
     "title": "Incorrect username or password",
-    "message": "Check that you have entered the correct username and password and try again"
+    "message": "Check that you have entered the correct username and password and try again."
   },
-  "additionalInformation": "In case you need to provide some additional info for the user"
+  "additionalInformation": "In case you need to provide some additional info for the user."
 }
